refactor: Rename and reorganize Jax distribution handlers module
