+++ tests/test_handlers_jax_dist.py
@@ -8,10 +8,9 @@ import jax
 import numpyro.distributions
 import pytest
 
-import effectful.handlers.jax.distribution as dist
 import effectful.handlers.jax.numpy as jnp
-from effectful.handlers.jax import jax_getitem, sizesof
-from effectful.handlers.jax.handlers import bind_dims, unbind_dims
+import effectful.handlers.numpyro as dist
+from effectful.handlers.jax import bind_dims, jax_getitem, sizesof, unbind_dims
 from effectful.ops.syntax import defop
 from effectful.ops.types import Operation, Term
